requirements.txt
browserbase>=0.3.0
playwright>=1.40.0
python-dotenv>=1.0.0
